requirements.txt
Django==1.5.1
Pillow==2.1.0
argparse==1.2.1
django-endless-pagination==2.0
django-yafinder==0.2
easy-thumbnails==1.3
wsgiref==0.1.2
